Rename files with references to lexicon
